feat: delete phase/group
